Refactor dependency management. Fixes #2081
Currently, the dependencies are managed with 2 methods: addModuleAsync to add a new module and removeModuleAsync to remove a module. The downside of this approach is that when you need to add/remove dependencies in bulk, you'll need to call those methods one by one. It's easy to introduce race conditions when doing this because every call to those methods will trigger a state update. Due to React's setState being async, the change is usually not propagated before the method is called again. This can have weird bugs, such as a removed dependency being added back again.

@nikki93 encountered this bug in #2081. It's actually pretty easy to repro, just remove a bunch of items from package.json at once and all of them but one will be added back.

The PR adds syncDependenciesAsync method to do the updates in one go. This method will take a map of dependencies and add/remove dependencies from existing list accordingly. It'll update the state only after all the operations are completed.

fbshipit-source-id: eb2112b